Reorganize the project file structure
