+++ static/konva.js
