--- step23/config.py
@@ -1,5 +1,0 @@
-JPOLY_WIDTH = 640
-JPOLY_HEIGHT = 480
-
-KONVA_WIDTH = 640
-KONVA_HEIGHT = 480
