--- step23/static/jPolygon.js
@@ -1,204 +1,0 @@
-/*
-   jPolygon - a ligthweigth javascript library to draw polygons over HTML5 canvas images.
-   Project URL: http://www.matteomattei.com/projects/jpolygon
-   Author: Matteo Mattei <matteo.mattei@gmail.com>
-   Version: 1.0
-   License: MIT License
-*/
-
-var perimeter = new Array();
-var complete = false;
-var canvas = document.getElementById("jPolygon");
-var ctx;
-
-function line_intersects(p0, p1, p2, p3) {
-    var s1_x, s1_y, s2_x, s2_y;
-    s1_x = p1['x'] - p0['x'];
-    s1_y = p1['y'] - p0['y'];
-    s2_x = p3['x'] - p2['x'];
-    s2_y = p3['y'] - p2['y'];
-
-    var s, t;
-    s = (-s1_y * (p0['x'] - p2['x']) + s1_x * (p0['y'] - p2['y'])) / (-s2_x * s1_y + s1_x * s2_y);
-    t = ( s2_x * (p0['y'] - p2['y']) - s2_y * (p0['x'] - p2['x'])) / (-s2_x * s1_y + s1_x * s2_y);
-
-    if (s >= 0 && s <= 1 && t >= 0 && t <= 1)
-    {
-        // Collision detected
-        return true;
-    }
-    return false; // No collision
-}
-
-function point(x, y){
-    ctx.fillStyle="white";
-    ctx.strokeStyle = "white";
-    ctx.fillRect(x-2,y-2,4,4);
-    ctx.moveTo(x,y);
-}
-
-function undo(){
-    ctx = undefined;
-    perimeter.pop();
-    complete = false;
-    start(true);
-}
-
-function clear_canvas(){
-    ctx = undefined;
-    perimeter = new Array();
-    complete = false;
-    document.getElementById('coordinates').value = '';
-    start();
-}
-
-function draw(end){
-    ctx.lineWidth = 1;
-    ctx.strokeStyle = "white";
-    ctx.lineCap = "square";
-    ctx.beginPath();
-
-    for(var i=0; i<perimeter.length; i++){
-        if(i==0){
-            ctx.moveTo(perimeter[i]['x'],perimeter[i]['y']);
-            end || point(perimeter[i]['x'],perimeter[i]['y']);
-        } else {
-            ctx.lineTo(perimeter[i]['x'],perimeter[i]['y']);
-            end || point(perimeter[i]['x'],perimeter[i]['y']);
-        }
-    }
-    if(end){
-        ctx.lineTo(perimeter[0]['x'],perimeter[0]['y']);
-        ctx.closePath();
-        ctx.fillStyle = 'rgba(255, 255, 255, 0.3)';
-        ctx.fill();
-        ctx.strokeStyle = 'white';
-        complete = true;
-    }
-    ctx.stroke();
-
-    // print coordinates
-    if(perimeter.length == 0){
-        document.getElementById('coordinates').value = '';
-    } else {
-        document.getElementById('coordinates').value = JSON.stringify(perimeter);
-    }
-}
-
-function check_intersect(x,y){
-    if(perimeter.length < 4){
-        return false;
-    }
-    var p0 = new Array();
-    var p1 = new Array();
-    var p2 = new Array();
-    var p3 = new Array();
-
-    p2['x'] = perimeter[perimeter.length-1]['x'];
-    p2['y'] = perimeter[perimeter.length-1]['y'];
-    p3['x'] = x;
-    p3['y'] = y;
-
-    for(var i=0; i<perimeter.length-1; i++){
-        p0['x'] = perimeter[i]['x'];
-        p0['y'] = perimeter[i]['y'];
-        p1['x'] = perimeter[i+1]['x'];
-        p1['y'] = perimeter[i+1]['y'];
-        if(p1['x'] == p2['x'] && p1['y'] == p2['y']){ continue; }
-        if(p0['x'] == p3['x'] && p0['y'] == p3['y']){ continue; }
-        if(line_intersects(p0,p1,p2,p3)==true){
-            return true;
-        }
-    }
-    return false;
-}
-
-function point_it(event) {
-    if(complete){
-        alert('Polygon already created');
-        return false;
-    }
-    var rect, x, y;
-
-    if(event.ctrlKey || event.which === 3 || event.button === 2){
-        if(perimeter.length==2){
-            alert('You need at least three points for a polygon');
-            return false;
-        }
-        x = perimeter[0]['x'];
-        y = perimeter[0]['y'];
-        if(check_intersect(x,y)){
-            alert('The line you are drowing intersect another line');
-            return false;
-        }
-        draw(true);
-        // alert('Polygon closed');
-	    event.preventDefault();
-        return false;
-    } else {
-        rect = canvas.getBoundingClientRect();
-        x = event.clientX - rect.left;
-        y = event.clientY - rect.top;
-        if (perimeter.length>0 && x == perimeter[perimeter.length-1]['x'] && y == perimeter[perimeter.length-1]['y']){
-            // same point - double click
-            return false;
-        }
-        if(check_intersect(x,y)){
-            alert('The line you are drowing intersect another line');
-            return false;
-        }
-        perimeter.push({'x':x,'y':y});
-        draw(false);
-        return false;
-    }
-}
-
-function start(with_draw) {
-    var img = new Image();
-    canvas.width = window.innerWidth * 0.35;
-    canvas.height = 300;
-
-    img.onload = function(){
-        var w = img.width;
-        var h = img.height;
-        if (img.width < canvas.width) {
-            canvas.width = img.width;
-            canvas.height = img.height;
-        } else {
-            h = h / w * canvas.width;
-            w = canvas.width;
-            canvas.height = h;
-        }
-        ctx = canvas.getContext("2d");
-        ctx.drawImage(img, 0, 0, w, h);
-        if(with_draw == true){
-            draw(false);
-        }
-    }
-    img.src = '/static/data/src.png';
-    
-}
-
-async function send_vertices() {
-    if (!complete) {
-        alert('Please close the selection!');
-        return;
-    }
-    var patch_size = [0, 0];
-    $.ajax({
-        async: false,
-        url: '/crop',
-        data: JSON.stringify({
-            "width": canvas.width, 
-            "height": canvas.height,
-            "perimeter": perimeter}),
-        type: "POST",
-        contentType: "application/json;charset=utf-8",
-        success: function(res){},
-        error: function(xhr, ajaxOptions, thrownError){
-            console.log(xhr.status);
-            console.log(thrownError);
-        }
-    });
-    return patch_size;
-}
