--- step23/templates/example.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-    <link rel="shortcut icon" href="{{ url_for('static', filename='favicon.ico') }}">
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-    <script src="https://unpkg.com/konva@8.0.4/konva.min.js"></script>
-    <head>
-        <title>Example using jPolygon</title>
-        <meta charset="UTF-8">
-        <meta name="description" content="An example using jPolygon">
-        <meta name="keywords" content="HTML,CSS,JavaScript,jPolygon">
-        <meta name="author" content="Matteo Mattei">
-    </head>
-    <body onload="clear_canvas()">
-        <div style="float:left">
-            <canvas id="jPolygon" style="cursor:crosshair" 
-                onmousedown="point_it(event)" oncontextmenu="return false;">
-                Your browser does not support the HTML5 canvas tag.
-            </canvas>
-        </div>
-        <div style="float:left; margin-left:20px;">
-            <div>
-                <button onclick="undo()">Undo</button>
-                <button onclick="clear_canvas()">Clear</button>
-                <button onclick="send_vertices(); start_konva()">Done</button>
-                <p>Press <strong>Left Click</strong> to draw a point.</p>
-                <p><strong>CTRL+Click</strong> or <strong>Right Click</strong> to close the polygon.</p>
-            </div>
-            <div>
-                <p><strong>Coordinates:</strong></p>
-                <textarea id="coordinates" disabled="disabled" style="width:300px; height:200px;"></textarea>
-            </div>
-        </div>
-        <div style="float: left">
-            <div id="konva"></div>
-            <button onclick="finish_transform()">Done</button>
-        </div>
-
-        <script type="text/javascript" src="{{ url_for('static', filename='jPolygon.js') }}"></script>
-        <script type="text/javascript" src="{{ url_for('static', filename='script.js') }}"></script>
-    </body>
-</html>
