--- web/static/data/.gitignore
@@ -1,2 +1,0 @@
-*.png
-*.jpg
